uicomponents: Move source files to internal directory
